更新门店地址信息时更新门店经纬度

+++ src/main/java/com/hy/salon/basic/service/ScheduleService.java
@@ -39,7 +39,10 @@ public class ScheduleService {
                 scheduleDao.deleteById(one.getRecordId());
             }
             if(schedule.getShiftId()==-1){
-                scheduleDao.deleteById(one.getRecordId());
+                if(one!=null){
+                    scheduleDao.deleteById(one.getRecordId());
+                }
+
             }
             scheduleDao.insert(schedule);
         }
